add flyway and finish user apis

+++ src/main/java/com/duy/thesisManagement/thesis/dto/UserDetailsImpl.java
@@ -13,14 +13,14 @@ import org.springframework.security.core.userdetails.UserDetails;
 public class UserDetailsImpl implements UserDetails {
 
 		private static final long serialVersionUID = 1L;
-		private Long id;
+		private Integer id;
 		private String username;
 		private String email;
 		@JsonIgnore
 		private String password;
 		private Collection<? extends GrantedAuthority> authorities;
 
-		public UserDetailsImpl(Long id, String username, String email, String password,
+		public UserDetailsImpl(Integer id, String username, String email, String password,
 				Collection<? extends GrantedAuthority> authorities) {
 				this.id = id;
 				this.username = username;
@@ -46,7 +46,7 @@ public class UserDetailsImpl implements UserDetails {
 				return authorities;
 		}
 
-		public Long getId() {
+		public Integer getId() {
 				return id;
 		}
